move write_to_file and unzip to downloader.py
